Fix invalid condition to add request to sender in Analyzer
It was "if data and self.sender", which, because of the lack of
__bool__ implementation, actually was "if data and len(self.sender)".
Which in turn was something like "if data and len(self.sender.queue)".
Which evaluated to False when the sender queue was empty and didn't
add Analyzer request.

tldr; kids, use "x is not None" if you actually only want to check
if something is None.